# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Kanban API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/boards": {
            "get": {
                "tags": [
                    "Boards"
                ],
                "summary": "Get list of all boards",
                "description": "Returns list of all boards",
                "operationId": "2a59923013089064327b5e1cd82b3559",
                "responses": {
                    "200": {
                        "description": "Array[] : contain all boards"
                    },
                    "419": {
                        "description": "Delay error : CSRF Token missed ?"
                    }
                }
            },
            "post": {
                "tags": [
                    "Boards"
                ],
                "summary": "Insert new board",
                "description": "Create new board to DB",
                "operationId": "7581840c73e5d20c98cec6eff906612e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "String -> modified board id"
                    },
                    "400": {
                        "description": "Bad request : No name || not connected"
                    },
                    "419": {
                        "description": "Delay error : CSRF Token missed ?"
                    }
                }
            }
        },
        "/api/boards/{id}": {
            "get": {
                "tags": [
                    "Boards"
                ],
                "summary": "Get info for specific board by id",
                "description": "Returns board by id",
                "operationId": "cf0bf9b1370bb623828295954dd93f0d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Array[] : Target board"
                    },
                    "419": {
                        "description": "Delay error : CSRF Token missed ?"
                    }
                }
            },
            "put": {
                "tags": [
                    "Boards"
                ],
                "summary": "Modify info for specific board by id",
                "description": "Modify board by id",
                "operationId": "512fd499e896a46d133bf76d6d37bd2e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of target board",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "is_archived": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "String -> modified board id"
                    },
                    "400": {
                        "description": "Bad request : No user || Not connected || Not correct rights"
                    },
                    "419": {
                        "description": "Delay error : CSRF Token missed ?"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Boards"
                ],
                "summary": "Delete board",
                "description": "Delete board by id",
                "operationId": "d1be3c718bb44cfeb4e38b698723dafe",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of target board",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Method not ready now"
                    }
                }
            }
        },
        "/api/boards/{id}/users": {
            "get": {
                "tags": [
                    "Boards"
                ],
                "summary": "Get users for specific board by id",
                "description": "Returns users by board id",
                "operationId": "785134fe84c530c9c8438fcd0a9ca03d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Array[] : Target board user's"
                    },
                    "400": {
                        "description": "Bad request : No user "
                    },
                    "419": {
                        "description": "Delay error : CSRF Token missed ?"
                    }
                }
            }
        },
...
    }
}